Move db sharing verification test into its own spec
- verification tests take a long time, separating into its own spec optimizes tests for Cypress parallelization
